refactor settings for compiler-timing transform: - add verifier support for each function analyzed and transformed - fix massive bug in wrapper pass dependencies --- ScalarEvolution   invoked last now - set canVectorize to false --- no expansion factor - set bottom guard threshold to be 0.5 * extension count - add loop debug information on ExtendLoop method, fix types for   LLS, instruction counts, fix constructor bug for setting loop   instruction counts

+++ src/llvm/compiler-timing/include/LatencyDFA.hpp
@@ -51,8 +51,8 @@ public:
     // ------- Analysis ------- 
     void ComputeDFA();
     set<Instruction *> *BuildIntervalsFromZero();
-    double GetLoopLatencySize() { return this->LoopLatencySize; }
-    double GetLoopInstructionCount() { return this->LoopInstructionCount; }
+    uint64_t GetLoopLatencySize() { return this->LoopLatencySize; }
+    uint64_t GetLoopInstructionCount() { return this->LoopInstructionCount; }
     double GetAccumulatedLatency(Instruction *I);
     double GetAccumulatedLatency(BasicBlock *BB);
     double GetIndividualLatency(Instruction *I);
